MoneyTime/api/views/users.py
from django.contrib.auth import get_user_model
from django_filters.rest_framework import DjangoFilterBackend
from rest_framework import filters
from rest_framework import viewsets
from rest_framework import permissions

from MoneyTime.api.serializers import UserSerializer


class UserViewSet(viewsets.ModelViewSet):
    """ViewSet for the User Model"""

    queryset = get_user_model().objects.all()
    serializer_class = UserSerializer
    permission_classes = [permissions.AllowAny]
    filter_backends = (
        filters.SearchFilter,
        filters.OrderingFilter,
        DjangoFilterBackend,
    )
