requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreschema==0.0.4
dj-database-url==0.5.0
Django==1.11.11
django-filter==1.1.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.2
psycopg2==2.7.4
pyparsing==2.2.0
python-dateutil==2.7.0
pytz==2018.3
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
